Identify pools by suffixes
Gitaly is using the repository's path to identify whether a repository
is a pool. Rails' pool paths are checked exhaustively with a regex and
Praefects paths are checked for a given prefix. These checks are not
sufficient to identify pools with transactions as the pool path will
be prefixed by the snapshot's root directory. Account for this by
changing the checks to verify the suffix of the path matches the
expected pool suffix.

A test was removed that is asserting behavior related to invalid pool
paths. Such state can't be created through the API anymore as we're
verifying the pool path in a stricter manner. We've never created
such paths as are being tested, so opted for removal rather than
maintaining the test.