Add keeparound entries to Transaction Manager
Some types of transactions, such as housekeeping, operate on a snapshot
repository. There is a gap between the transaction doing its work and
the time when it is committed. They need to verify if concurrent
operations can cause conflict. This commit lets the transaction manager
maintain a list of keeparound log entries. These log entries are still
kept around even after they are applied. They are removed when no active
readers are accessing the corresponding snapshots.