procreceive: Introduce and add ProcReceiveRegistry
Introduce and add:
1. ProcReceiveRegistry
2. ProcReceiveHookInvocation

ProcReceiveRegistry is the registry which provides the proc-receive hook
invocation mechanism against a provided transaction ID. The registry
allows RPCs to communicate with the git-proc-receive hook and receive
information about the reference updates to be performed, then the RPCs
can interact with the transaction manager and accept or reject reference
updates, this information is relayed to the proc-receive hook which will
relay the information to the user.

ProcReceiveHookInvocation is an interface which provides abstraction
around the proc-receive invocation provided by the ProcReceiveRegistry.
The interface allows the users to obtain reference updates, meta
information around these updates and functions to accept or reject
individual updates.