hook: Add `ProcReceiveHook` to `GitLabHookManager`
Add a new function `ProcReceiveHook` to `GitLabHookManager`. The
`ProcReceiveHook` is used to intercept git-receive-pack(1)'s
execute-commands code. This allows us to intercept the reference updates
and avoid writing directly to the disk. The intercepted updates are then
bundled into `procReceiveHookInvocation` and added to the registry. The
RPC which invoked git-receive-pack(1) in the first place picks up the
invocation from the RPC and accepts/rejects individual references.

We also add the `ProcReceiveRegistry` to the `GitLabHookManager` so that
the `ProcReceiveHook` has access to the registry.