Extract helpers and executor out of TestTransactionManager
Recently, the test TestTransactionManager has been quite long with
nearly ~6000 LOCs. All test cases are packed into one big test table.
That makes it hard to navigate through the file. The size of the file
even breaks GitLab.com's review feature. This commit restructures the
test file so that we can split it into smaller ones. The foundational
changes consist of the following:
- Move common test structs to test helper file.
- Generalize and extract test setup.
- Extract scenario execution.