Allow configuring target repo for rails-specs
In some situations we will want to run the rails-specs job against
a repository other than the default gitlab-org/gitlab. Examples of
this may be against security release.

We can allow a default value canbe set, but when the need arises,
an ENV can be set to target the job against the relevant repo.

Fixes https://gitlab.com/gitlab-org/gitaly/-/issues/5727