backup: Embed serverRepository in restore struct
Refactors the restoreRequest struct to embed the existing
serverRepository struct instead of redefining the same fields. This is
done as backups and restore jobs share the same inputs except for the
AlwaysCreate option.

Additionally, this is similar to how the structure is defined in Rails,
see: https://gitlab.com/gitlab-org/gitlab/-/blob/master/lib/backup/gitaly_backup.rb