Integrate log exporter hook with command factory
This commit will introduce a new feature flag DangerousForceCollectAllTraces. If enabled, git trace2 API events will be logged in Gitaly node logs as well. This is all performed in a new hook as part of DefaultTrace2Hooks. This hook's handler also uses a rate limiter from the golang.org/x/time rate package utilizing a token bucket algorithm. It uses .Reserve() func to determine if the invocation is permitted and not exceeding the rate limit. It will return an error if the rate limit is exceeded and the trace will be skipped.