gitattributes: Get rid of endOfAttributes
`gitattributes.CheckAttr` appended "end-of-attributes-----" to the names
of the attributes it wants to read. When the output of git-check-attr(1)
is parsed, Gitaly reads until this attribute is found.

Instead, memorize the number of attributes the user requested and keep
reading until that number of attributes is read. This works because the
user should not do concurrent calls to `Check()`

Label: maintenance::refactor