hook: Propagate error's via `ProcReceiveHandler`
Currently the `Close()` function of the `ProcReceiveHandler` only
provides information on whether the handler's job is completed or not.
But it would be useful for the handler to also propagate errors, this
could be used by the ProcReceive server to return the appropriate exit
code to git-receive-pack(1).

So let's convert the `Close()` function to accept an error and propagate
this error via the `doneCh`.