backup: Stop sending invalid requests
Modifies the Restore CLI tests so we stop sending an invalid restore
command as the final JSON object into stdin. This is required as a
subsequent commit will move the repository removal logic to execute
after the Pipeline completes successfully. If the tests purposely cause
the pipeline to fail, the restore logic will never execute.

Coverage of the Pipeline's error messaging is covered separately in the
Pipeline's unit tests.