backup: Only remove "dangling" repositories
Instead of invoking the RemoveAll() RPC prior to the restore, we instead
compare the set of existing repositories to the set of repositories
contained in the backup being restored. The difference between the two
sets -- the set of "dangling" repositories -- are removed individually.

This is done to ensure the state of repos in Gitaly matches the
worldview held by the Rails DB after a GitLab instance restore.