hook: Use `context.Context` in `ProcReceiveRegistry.Transmit`
The `Transmit()` function provided by `ProcReceiveRegistry` is a
blocking function, which ensures that a handler is transmitted to any
waiters.

The problem is, that there is a potential scenario where the waiter
might have exited early. Hence blocking this call forever. So to avoid
that, we expand the function to receive a context. With this, if the
main RPC (waiter) exits early, the context will be canceled and we wont
be blocked forever.