Defualt enable CLONE_INTO_CGROUP support
CLONE_INTO_CGROUP support is currently default disabled. We can't test
it on production as we're not running cgroup v2. Cgroup v2 is also
broken when used through Omnibus as Gitaly doesn't run in the same
hierarchy as the child cgroups and can't thus move move the processes
it spawns into those cgroups.

To remove the feature flag before we support cgroup v2 ourselves, let's
just enable the flag by default and remove it a couple of releases after
if there are no issues reported. Manual testing has shown it to work.
Default enabling this would only change behavior for setups that manually
configure the cgroups and avoid the issue with moving processes across
hierarchies.

Changelog: fixed