backup: Extract bundle restore into own function
Extracts the loop responsible for restoring each step of a backup into
a separate function, along with the call to remove the repository. As a
consequence of the latter change, the tests have been updated to expect
the repository to exist even if the backup does not.

This change is done to facilitate a subsequent MR which will introduce
another function that optimises the restore process by resetting the ref
list rather than using bundles.